requirements.txt
streamlit==1.35.0
requests
Pillow
contextily
matplotlib
pyproj
numpy
scikit-image
requests
bs4